Correctly render mentions in events. Extract shared specs and logic for comment events. Closes #5177
