+++ decidim-comments/lib/decidim/comments/test.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 require "decidim/comments/test/shared_examples/create_comment_context"
+require "decidim/comments/test/shared_examples/comment_event"
