+++ decidim-comments/spec/events/decidim/comments/comment_by_followed_user_event_spec.rb
@@ -5,22 +5,12 @@ require "spec_helper"
 module Decidim
   module Comments
     describe CommentByFollowedUserEvent do
-      include Decidim::ComponentPathHelper
-      include Decidim::SanitizeHelper
-
-      include_context "when a simple event"
+      include_context "comment event"
+      it_behaves_like "a comment event"
 
       let(:event_name) { "decidim.events.comments.comment_by_followed_user" }
-      let(:comment) { create(:comment) }
       let(:resource) { comment.root_commentable }
-      let(:resource_title) { decidim_html_escape resource.title }
-      let(:extra) { { comment_id: comment.id } }
       let(:author) { comment.author }
-      let(:author_name) { decidim_html_escape author.name }
-      let(:author_path) { author_presenter&.profile_path.to_s }
-      let(:author_nickname) { author_presenter&.nickname.to_s }
-
-      it_behaves_like "a simple event"
 
       describe "email_subject" do
         it "is generated correctly" do
@@ -48,12 +38,6 @@ module Decidim
             .to start_with("There is a new comment by <a href=\"#{author_path}\">#{author_name} #{author_nickname}</a> in")
           expect(subject.notification_title)
             .to end_with("<a href=\"#{resource_path}#comment_#{comment.id}\">#{resource_title}</a>.")
-        end
-      end
-
-      describe "resource_text" do
-        it "outputs the comment body" do
-          expect(subject.resource_text).to eq comment.body
         end
       end
     end
