+++ decidim-comments/spec/events/decidim/comments/reply_created_event_spec.rb
@@ -3,19 +3,13 @@
 require "spec_helper"
 
 describe Decidim::Comments::ReplyCreatedEvent do
-  include Decidim::SanitizeHelper
-  include_context "when a simple event"
+  include_context "comment event"
+  it_behaves_like "a comment event"
 
   let(:resource) { comment.root_commentable }
-  let(:resource_title) { comment.root_commentable.title }
   let(:parent_comment) { create :comment }
   let(:comment) { create :comment, commentable: parent_comment, root_commentable: parent_comment.root_commentable }
-  let(:comment_author) { comment.author }
-  let(:comment_author_name) { decidim_html_escape comment.author.name }
   let(:event_name) { "decidim.events.comments.reply_created" }
-  let(:extra) { { comment_id: comment.id } }
-
-  it_behaves_like "a simple event"
 
   describe "email_subject" do
     it "is generated correctly" do
@@ -43,12 +37,6 @@ describe Decidim::Comments::ReplyCreatedEvent do
 
       expect(subject.notification_title)
         .to end_with("your comment in <a href=\"#{resource_path}#comment_#{comment.id}\">#{resource.title}</a>")
-    end
-  end
-
-  describe "resource_text" do
-    it "outputs the comment body" do
-      expect(subject.resource_text).to eq comment.body
     end
   end
 end
