+++ decidim-comments/spec/events/decidim/comments/user_mentioned_event_spec.rb
@@ -3,15 +3,17 @@
 require "spec_helper"
 
 describe Decidim::Comments::UserMentionedEvent do
-  include_context "when a simple event"
-
-  let(:resource) { comment.commentable }
-  let(:comment) { create :comment }
-  let(:comment_author) { comment.author }
   let(:event_name) { "decidim.events.comments.user_mentioned" }
-  let(:extra) { { comment_id: comment.id } }
 
-  it_behaves_like "a simple event"
+  include_context "comment event"
+  it_behaves_like "a comment event"
+
+  before do
+    body = "Comment mentioning some user, @#{comment.author.nickname}"
+    parsed_body = Decidim::ContentProcessor.parse(body, current_organization: comment.organization)
+    comment.body = parsed_body.rewrite
+    comment.save
+  end
 
   describe "email_subject" do
     it "is generated correctly" do
@@ -41,10 +43,10 @@ describe Decidim::Comments::UserMentionedEvent do
         .to include(" by <a href=\"/profiles/#{comment_author.nickname}\">#{comment_author.name} @#{comment_author.nickname}</a>")
     end
   end
-
   describe "resource_text" do
-    it "outputs the comment body" do
-      expect(subject.resource_text).to eq comment.body
+    it "correctly renders comments with mentions" do
+      expect(subject.resource_text).not_to include("gid://")
+      expect(subject.resource_text).to include("@#{comment.author.nickname}")
     end
   end
 end
